+++ src/main/java/io/github/newhoo/restkit/common/RestClientData.java
@@ -15,6 +15,7 @@ public class RestClientData {
 
     private String url;
     private String method;
+    private Map<String, String> config;
     private Map<String, String> headers;
     private Map<String, String> params;
     private String body;