+++ src/main/java/io/github/newhoo/restkit/common/RestModule.java
@@ -13,6 +13,10 @@ public class RestModule {
 
     private List<RestItem> restItems;
 
+    public String getOrderKey() {
+        return restItems.isEmpty() ? "" : restItems.get(0).getFramework();
+    }
+
     @Override
     public String toString() {
         return this.moduleName;