+++ src/main/java/io/github/newhoo/restkit/toolwindow/RestServiceTree.java
@@ -48,6 +48,7 @@ import java.awt.*;
 import java.awt.event.InputEvent;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -238,6 +239,7 @@ public class RestServiceTree extends JPanel implements DataProvider {
                              .entrySet()
                              .stream()
                              .map(entry -> new RestModule(entry.getKey(), entry.getValue()))
+                             .sorted(Comparator.comparing(RestModule::getOrderKey))
                              .collect(Collectors.toList())
         );
     }
@@ -423,10 +425,10 @@ public class RestServiceTree extends JPanel implements DataProvider {
                     description = split[2];
                 }
             }
-            String tooltip = "protocol: " + myRestItem.getProtocol() + "<br/>"
-                    + "url: " + myRestItem.getUrl() + "<br/>"
-                    + "desc: " + description + "<br/>"
-                    + "from: " + myRestItem.getFramework();
+            String tooltip = "P: " + myRestItem.getProtocol() + "<br/>"
+                    + "U: " + myRestItem.getUrl() + "<br/>"
+                    + "D: " + description + "<br/>"
+                    + "F: " + myRestItem.getFramework();
             getTemplatePresentation().setTooltip(tooltip);
         }